DisableEventIndex: Rename the file to contain the Dialog suffix.
